Remove obsoletes directive
The obsoletes directive should, according to the documentation, only be
used when a package changes name or a new package obsoletes an old
package. For normal versioning it shouldn't be used and it is making it
impossible to upgrade the rpm packages we build as the newer package
will be rejected with an error saying it's obsolete. Removing the
directive seems to fix this issue.